Add files via upload
Section where the simulation objects are configured, such as the 6-DOF model, reading of the coefficients used, main plotting routine, etc. Note the need to change the coefficient inputs if you intend to use this code. The coefficients may also be defined differently depending on the application, which requires care and corresponding adjustments in the code.